moved shared and publish to new shared @ folder
