+++ imports/api/methods/work-methods.js
@@ -1,5 +1,5 @@
 import { Meteor } from 'meteor/meteor';
-import { getHost } from './shared';
+import { getHost } from '../@/shared';
 
 Meteor.methods({
   getAllWorks() {
